code/requirements.txt
scikit-learn
pandas
np
matplotlib
PyYAML
google.generativeai
streamlit